icatchup

+++ SCRATCH_ZONE_20/in_Tcl_Module/tcl-parse-v2/tcl_parser_library.cxx
@@ -79,8 +79,10 @@ int tokeniser_init_restricted( Tcl_Interp * interp_client )
 
 	// module_tokeniser * mod = new( object );
 	int mod = 9876;
-	ClientData cmd_obj = (ClientData) mod;
-	printf("sizeof int %d sizeof ClientData %d \n", sizeof(int), sizeof(ClientData) );
+	ClientData cmd_obj = (ClientData) (long long) mod; // cast from 32 to 64
+	// C is stupid, sizoeof(word) is 64 not 32 bit
+	// for structs larger than 2G
+	printf("sizeof int %ld sizeof ClientData %ld \n", sizeof(int), sizeof(ClientData) );
 	Tcl_ObjCmdProc * cmd_fn = tokeniser_new;
 	Tcl_CmdDeleteProc * cmd_del = NULL;
